SIMPLY-3644 Refactor bookmark creation from server annotation + R2 location
